refactor: create hero components
